some common methods moved from ModelClassesToTsInterfacesConverter to ModelClassesAbstractConverter.java
