+++ spring-rest2ts-generator/src/main/java/com/blueveery/springrest2ts/converters/ModelClassesToTsInterfacesConverter.java
@@ -89,58 +89,4 @@ public class ModelClassesToTsInterfacesConverter extends ModelClassesAbstractCon
 
     }
 
-    private void setAsNullableType(Property property, TSField tsField, NullableTypesStrategy nullableTypesStrategy) {
-        if (property.getGetterType() != null) {
-            nullableTypesStrategy.setAsNullableType(property.getGetterType(), property.getDeclaredAnnotations(), tsField);
-            return;
-        }
-
-        if (property.getSetterType() != null && Objects.equals(property.getGetterType(), property.getSetterType())) {
-            nullableTypesStrategy.setAsNullableType(property.getSetterType(), property.getDeclaredAnnotations(), tsField);
-        }
-    }
-
-    private SortedSet<Property> getClassProperties(Class javaClass, ObjectMapper objectMapper) {
-        Map<String, Property> propertyMap = new HashMap<>();
-        int currentIndex = 0;
-
-        for (Field field : javaClass.getDeclaredFields()) {
-            if (!Modifier.isTransient(field.getModifiers()) && !Modifier.isStatic(field.getModifiers())) {
-                if(objectMapper.filter(field)) {
-                    Property property = new Property(objectMapper.getPropertyName(field), currentIndex++, field);
-                    propertyMap.put(property.getName(), property);
-                }
-            }
-        }
-
-        for (Method method : javaClass.getDeclaredMethods()) {
-            if (!Modifier.isStatic(method.getModifiers())) {
-                if(couldBeGetter(method) && objectMapper.filter(method, true)){
-                    String propertyName = objectMapper.getPropertyName(method, true);
-                    final int newIndex = currentIndex++;
-                    Property property = propertyMap.computeIfAbsent(propertyName, (key) -> new Property(key, newIndex));
-                    property.setGetter(method);
-                }
-
-                if(couldBeSetter(method) && objectMapper.filter(method, false)){
-                    String propertyName = objectMapper.getPropertyName(method, false);
-                    final int newIndex = currentIndex++;
-                    Property property = propertyMap.computeIfAbsent(propertyName, (key) -> new Property(key, newIndex));
-                    property.setSetter(method);
-                }
-
-            }
-        }
-
-        return new TreeSet<>(propertyMap.values());
-    }
-
-    private boolean couldBeGetter(Method method) {
-        return method.getParameterCount() == 0 && method.getReturnType() != void.class;
-    }
-
-    private boolean couldBeSetter(Method method) {
-        return method.getParameterCount() == 1 && method.getReturnType() == void.class;
-    }
-
 }
